requirements.txt
djangorestframework
python-dotenv
requests
django
django-phonenumber-field
phonenumbers
django-extensions
bpython
